requirements.txt
pypokedex
pillow
urllib3